+++ src/App.js
@@ -1,5 +1,5 @@
 import Header from "./components/Header/Header";
-import MovieList from "./components/MovieList";
+import MovieList from "./components/MovieList/MovieList";
 
 class App {
   $target;
